Usecase classes -- refactors use cases and notices
