+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidEmailNotice.java
@@ -44,10 +44,6 @@ public class InvalidEmailNotice extends ErrorNotice {
         return emailValue;
     }
 
-    public String getEntityId() {
-        return entityId;
-    }
-
     public String getFieldName() {
         return fieldName;
     }
