+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidLangNotice.java
@@ -45,10 +45,6 @@ public class InvalidLangNotice extends ErrorNotice {
         return langValue;
     }
 
-    public String getEntityId() {
-        return entityId;
-    }
-
     public String getFieldName() {
         return fieldName;
     }
