+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/MissingShortAndLongNameForRouteNotice.java
@@ -35,8 +35,4 @@ public class MissingShortAndLongNameForRouteNotice extends ErrorNotice {
     public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
-
-    public String getEntityId() {
-        return entityId;
-    }
 }
