+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/warning/MissingRouteLongNameNotice.java
@@ -34,9 +34,5 @@ public class MissingRouteLongNameNotice extends WarningNotice {
     public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
-
-    public String getEntityId() {
-        return entityId;
-    }
 }
