+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/warning/MissingRouteShortNameNotice.java
@@ -34,8 +34,4 @@ public class MissingRouteShortNameNotice extends WarningNotice {
     public void export(final NoticeExporter exporter) throws IOException {
         exporter.export(this);
     }
-
-    public String getEntityId() {
-        return entityId;
-    }
 }
